+++ indico/modules/events/controllers/creation.py
@@ -46,13 +46,15 @@ class RHCreateEvent(RHProtected):
 
     def _process_args(self):
         self.event_type = EventType[request.view_args['event_type']]
+        self.root_category = Category.get_root()
+        self.single_category = not self.root_category.children
 
     @cached_property
     def _default_category(self):
         try:
             category_id = int(request.args['category_id'])
         except (ValueError, KeyError):
-            return None
+            return self.root_category if self.single_category else None
         else:
             return Category.get(category_id, is_deleted=False)
 
@@ -122,4 +124,4 @@ class RHCreateEvent(RHProtected):
                 notify_event_creation(event)
             return jsonify_data(flash=False, redirect=url_for('event_management.settings', event))
         return jsonify_template('events/forms/event_creation_form.html', form=form, fields=form._field_order,
-                                event_type=self.event_type.name)
+                                event_type=self.event_type.name, single_category=self.single_category)